Split ambiguity detection logic off into its own file
